refactor: extract theme color parser
